Create BotRollMasterLoot.cpp
This action allows bots to respond automatically to a "roll <itemlink>" command in the raid chat when the player is using MasterLoot mode. 

It simulates a /rand roll as a real player would do in-game and allows the Master Looter to see the results of the roll for proper loot distribution.